+++ source/library/Language/Haskell/Brittany/Internal/Layouters/Decl.hs
@@ -78,6 +78,8 @@ layoutDeclWithExactText exactText hasSourceComments d@(L loc decl) = case decl o
   TyClD _ tycl
     | requiresExactTypeDeclaration tycl || requiresExactTypes d ->
         layoutExact TypeClassDeclarationFallback d exactText
+    | supportsCommentedDataDecl tycl ->
+        withTransformedAnns d $ layoutTyCl (L loc tycl)
     | otherwise -> layoutExactWhenCommented d $ layoutTyCl (L loc tycl)
   InstD _ (TyFamInstD _ tfid) ->
     layoutExactWhenCommented d $ layoutTyFamInstDecl False d tfid